+++ app.js
@@ -5,10 +5,8 @@ const mongoose = require("mongoose")
 const session = require("express-session")
 const flash = require("connect-flash")
 const passport = require("passport")
-const bcrypt = require("bcryptjs")
+const userRoute = require("./routes/user")
 
-//models
-const User = require("./model/User");
 
 // Specifing app port
 const PORT = 8070;
@@ -39,6 +37,7 @@ app.use(session({
     resave: true,
     saveUninitialized: false,
 }));
+
 
 // express messages middleware
 app.use(require('connect-flash')());
@@ -75,10 +74,7 @@ app.get("*",(request,response,next)=>{
     next();
 })
 
-
-/*
-*    Request routes
-*/
+app.use("/users",userRoute);
 
 // Get request to the starter page
 app.get("/",(request,response)=>{
@@ -86,113 +82,6 @@ app.get("/",(request,response)=>{
         title: "Home page"
     });
     return;
-});
-
-// User list route
-app.get('/users/', (request,response)=>{
-    User.find({},(err,users)=>{
-        if(err){
-            console.log(err)
-        }else{
-            response.render("Users",{
-                users : users,
-                title : "User list"
-            });
-            return;
-        }
-    })
-});
-
-// Adding route files
-app.get("/users/add",(request,response)=>{
-    response.render("AddUser",{
-        title : "Add user"
-    });
-    return;
-});
-app.post('/users/add',(request,response) =>{
-    const name = request.body.name;
-    const email = request.body.email;
-    const username = request.body.username;
-    const password = request.body.password;
-    const passwordSalt = request.body.passwordSalt;
-
-    bcrypt.genSalt(10,(err,salt)=>{
-        bcrypt.hash(request.body.password,salt,(err,hash)=>{
-            if(err){
-                console.log(err);
-            }
-            let model = new User({
-                name : name,
-                email : email,
-                username : username,
-                password : hash,
-                dateCreated : Date.now(),
-                isDeleted : false,
-            });
-            model.save((err)=>{
-                if(err){
-                    console.log(err);
-                    request.flash("error", "User failed to add successfully.")
-                }
-                else{
-                    console.log("user added successfully.");
-                    request.flash("success", "User added successfully.")
-                    response.redirect("/users/")
-                }
-            });
-        });
-    })
-
-});
-
-app.get("/users/edit/:id",(request,response)=>{
-    User.findById(request.params.id,(err,user)=>{
-        response.render("EditUser",{
-            user : user,
-            title : "Edit user"
-        });
-        return;
-    });
-});
-
-app.post("/users/edit/:id",(request,response)=>{
-    const username = request.body.username;
-    const email = request.body.email;
-    const name = request.body.name;
-    const id = request.params.id;
-
-    let query = {_id : id};
-
-    let model = {
-        username : username,
-        name : name,
-        email : email
-    };
-    User.update(query,model,(err)=>{
-        if(err){
-            console.log(err)
-            request.flash("error", "Save was unsuccessful.")
-        }else{
-            request.flash("success", "User updated successfully.")
-            response.redirect("/users")
-        }
-    })
-})
-
-app.delete("/users/delete/:id",(request,response)=>{
-    let query = {_id : request.params.id};
-     User.findById(request.params.id,(err, user)=>{
-        User.remove(query,(err)=>{
-            if(err){
-                console.log(err);
-            }
-            else{
-                response.send("success");
-                request.flash("success", "User deleted successfully.")
-            }
-        });
-     });
 });
 
 app.get("/account/login",(request,response)=>{
